Деактивация/активация формы перемещена в init.js. Реализован пункт ТЗ о запрете отправки невалидных данных.
