+++ js/main.js
@@ -1,9 +1,11 @@
-import {generateAdsHTML} from './modules/generate-ads-html.js';
-import {disableForms, enableForms} from './modules/form.js';
+import { generateAdsHTML } from './modules/generate-ads-html.js';
+import { disableForms, enableForms } from './modules/init.js';
+import { replaceSubmitHandler } from './modules/forms.js';
 
 
 window.addEventListener('DOMContentLoaded', () => {
   generateAdsHTML(3);
+  replaceSubmitHandler('.ad-form');
 
   // disableForms();
   // Для проверки работоспособности функции включения форм
